start on more memory friendly version of command queue
This avoids the pointers (and therefore using slice allocations w/ reuse)
by keeping fixed sized structs for most things and multiple arrays for
the in-flight changes.

To avoid memmove of batches we have a next_batch_index int that can be
changed as we append batches to the end of the array to reorder.

I'm still thinking about what we can do to track rects as we do reordering
as well as what operations are barriers. Will probably keep a window by
program as we go which gets reset as we hit barriers.